pyproject.toml
[project]
name = "query-tc-unicode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "pydoll-python>=2.15.1",
]
